[maven-release-plugin]  copy for tag apache-mime4j-0.5
git-svn-id: https://svn.apache.org/repos/asf/james/mime4j/tags/apache-mime4j-0.5@703860 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/james/mime4j/message/TempFileTextBody.java
@@ -104,9 +104,6 @@ class TempFileTextBody extends AbstractBody implements TextBody {
      * @see org.apache.james.mime4j.message.Body#writeTo(java.io.OutputStream, int)
      */
     public void writeTo(OutputStream out, int mode) throws IOException {
-        if (disposed)
-            throw new IllegalStateException("TempFileTextBody has been disposed");
-
         final InputStream inputStream = tempFile.getInputStream();
         CodecUtil.copy(inputStream, out);
     }
@@ -117,13 +114,10 @@ class TempFileTextBody extends AbstractBody implements TextBody {
      * @see org.apache.james.mime4j.message.Disposable#dispose()
      */
     public void dispose() {
-        try {
+        if (tempFile != null) {
             tempFile.delete();
-        } finally {
-            mimeCharset = null;
             tempFile = null;
-
-            super.dispose();
         }
     }
+
 }
